+++ selfdrive/ui/qt/onroad.cc
@@ -1,7 +1,5 @@
 #include "selfdrive/ui/qt/onroad.h"
 
-#include <iostream>
-
 #include "selfdrive/common/swaglog.h"
 #include "selfdrive/common/timing.h"
 #include "selfdrive/ui/paint.h"
@@ -16,8 +14,9 @@ OnroadWindow::OnroadWindow(QWidget *parent) : QWidget(parent) {
   layout->setStackingMode(QStackedLayout::StackAll);
 
   // old UI on bottom
-  nvg = new NvgWindow(this);
-  QObject::connect(this, &OnroadWindow::update, nvg, &NvgWindow::update);
+  nvg = new RoadView(this);
+  layout->addWidget(nvg);
+  QObject::connect(this, &OnroadWindow::update, nvg, &RoadView::update);
 
   QHBoxLayout* split = new QHBoxLayout();
   split->setContentsMargins(0, 0, 0, 0);
@@ -187,49 +186,4 @@ void OnroadAlerts::paintEvent(QPaintEvent *event) {
     configFont(p, "Open Sans", 88, "Regular");
     p.drawText(QRect(0, r.height() - (l ? 361 : 420), width(), 300), Qt::AlignHCenter | Qt::TextWordWrap, text2);
   }
-}
-
-
-NvgWindow::NvgWindow(QWidget *parent) : QOpenGLWidget(parent) {
-  setAttribute(Qt::WA_OpaquePaintEvent);
-}
-
-NvgWindow::~NvgWindow() {
-  makeCurrent();
-  doneCurrent();
-}
-
-void NvgWindow::initializeGL() {
-  initializeOpenGLFunctions();
-  std::cout << "OpenGL version: " << glGetString(GL_VERSION) << std::endl;
-  std::cout << "OpenGL vendor: " << glGetString(GL_VENDOR) << std::endl;
-  std::cout << "OpenGL renderer: " << glGetString(GL_RENDERER) << std::endl;
-  std::cout << "OpenGL language version: " << glGetString(GL_SHADING_LANGUAGE_VERSION) << std::endl;
-
-  ui_nvg_init(&QUIState::ui_state);
-  prev_draw_t = millis_since_boot();
-}
-
-void NvgWindow::update(const UIState &s) {
-  // Connecting to visionIPC requires opengl to be current
-  if (s.vipc_client->connected){
-    makeCurrent();
-  }
-  repaint();
-}
-
-void NvgWindow::resizeGL(int w, int h) {
-  ui_resize(&QUIState::ui_state, w, h);
-}
-
-void NvgWindow::paintGL() {
-  ui_draw(&QUIState::ui_state, width(), height());
-
-  double cur_draw_t = millis_since_boot();
-  double dt = cur_draw_t - prev_draw_t;
-  if (dt > 66 && !QUIState::ui_state.scene.driver_view) {
-    // warn on sub 15fps
-    LOGW("slow frame time: %.2f", dt);
-  }
-  prev_draw_t = cur_draw_t;
 }
